requirements.txt
requests
pandas
numpy
matplotlib
psycopg2-binary
apache-airflow
pyspark